Builds on Debian 12 with gcc symlinked to gcc-9
Install in /usr/local/lib/ptxdist-2020.08.0/bin:
gcc -> /usr/bin/gcc-9
gcc-ar -> /usr/bin/gcc-ar-9
gcc-nm -> /usr/bin/gcc-nm-9
gcc-ranlib -> /usr/bin/gcc-ranlib-9

+++ src/liboslinux/liboslinux/OsThread.c
@@ -169,7 +169,7 @@ static i32 OsThread_Attributes(pthread_attr_t* attribute,
          break;
       }
 
-      if (stackSize <= PTHREAD_STACK_MIN)
+      if (stackSize <= (size_t)PTHREAD_STACK_MIN)
       {
          stackSize = 2*PTHREAD_STACK_MIN;
       }
